Casts and other minor mods to avoid warnings

+++ Source Code/GraspVR/GraspTrackers.cpp
@@ -165,7 +165,7 @@ unsigned int GraspDexTrackers::GetTrackerStatus( void ) {
 			}
 			if ( visibility ) group_count++;
 		}
-		status = status + pow(10.0,group) * group_count;
+		status = status + (int) pow(10.0,group) * group_count;
 	}
 	return( status );
 }
